refactor

+++ src/main/java/com/rbkmoney/registry/payout/worker/model/Payouts.java
@@ -7,10 +7,10 @@ import java.util.Map;
 
 @Data
 public class Payouts {
-    public Map<String, Map<String, Float>> payouts = new HashMap<>();
+    public Map<String, Map<String, Long>> payouts = new HashMap<>();
 
-    public void putAll(Map<String, Map<String, Float>> invoicePayment,
-                       Map<String, Map<String, Float>> invoiceRefund) {
+    public void putAll(Map<String, Map<String, Long>> invoicePayment,
+                       Map<String, Map<String, Long>> invoiceRefund) {
         payouts.putAll(invoicePayment);
         payouts.putAll(invoiceRefund);
     }
