+++ src/main/java/com/rbkmoney/registry/payout/worker/model/Transactions.java
@@ -6,8 +6,8 @@ import org.springframework.util.MultiValueMap;
 
 @Data
 public class Transactions {
-    public MultiValueMap<String, Float> invoicePayments = new LinkedMultiValueMap<>();
-    public MultiValueMap<String, Float> invoiceRefunds = new LinkedMultiValueMap<>();
+    public MultiValueMap<String, Long> invoicePayments = new LinkedMultiValueMap<>();
+    public MultiValueMap<String, Long> invoiceRefunds = new LinkedMultiValueMap<>();
 
     public void addAll(Transactions transactions) {
         if (transactions.getInvoicePayments() != null) {
